requirements.txt
flask>=3.0
reportlab>=4.0
gunicorn>=21.0
pypdf>=4.0
pillow>=10.0
pywebview>=5.0
